PayTrace: Support level_3_data fields
PayTrace supports the addition of level_3_data to any approved and unsettled sale transaction. In order to add level_3_data to an approved transaction, the transaction will have to have been completed and returned a transaction_id. Due to this workflow, the method to add level_3_data has been built into the end of the purchase method. As of now, only Visa and MasterCard are supported for level_3_data so an additional test has been included in the remote file to make sure that if a card other than visa or MasterCard is processed, the purchase will go through and the attempt to add level_3_data will be ignored.

CE-1737

Local:
4775 tests, 73690 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed
Unit:
18 tests, 70 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed
Remote:
24 tests, 58 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed