Worldpay: Synchronous Response with purchase transaction
Worldpay is moving to a new synchronous response workflow and in some
scenarios will lead to different success cases, particularly for the
`purchase` transaction, where a `MultiResponse`, asynchronous
transaction workflow is invoked.

For some MID configurations, a transaction can be executed with a
`skip_capture` option, which allows for the `authorize` to be
executed and subsequently will be implicitly captured, rendering a
`CAPTURED` value for the `lastEvent` tag.

This change is for broadening the scope of our success criteria within
the context of an `authorize` transaction to allow for `AUTHORISED` or
`CAPTURED`.

CE-1786

Unit: 89 tests, 542 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote: 71 tests, 304 assertions, 2 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
97.1831% passed