Credorax: Determine ISK decimal by datetime
Credorax is changing the decimal value for the ISK currency on August
31 2021 8:00 UTC. This determines how the amount should be calculated
by checking against this deadline.

Their sandbox already considers ISK zero-decimal, so it is difficult to
test a behavior that is dependent on whether you're running transactions
against the sandbox or a live endpoint, so I have emulated the change
on the sandbox by setting a date in the past, in the same way as the
future deadline is set for Live.

Remote:
45 tests, 164 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Unit:
79 tests, 374 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed