EBANX: Added `processing_type` Gateway Specific Field
The headers that are generated for the `commit` method are now generated via a helper method. This method will look for a gateway specific field `processing_type`; if the field is present with a value of `local`, an additional header will be added to the gateway request: `x-ebanx-api-processing-type`.

CE-2113

Rubocop:
725 files inspected, no offenses detected

Local:
4995 tests, 74785 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
Loaded suite test/remote/gateways/remote_ebanx_test
25 tests, 69 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed