Braintree: Add ACH support for store
Summary:
------------------------------

This commit modifies the store method to add support for bank accounts
in order to enable ACH transactions on purchases, for that is needed an
extra work to simulate the tokenization flow that Braintree enables for
the front-end libraries to create token-nonces for payment methods.

This nonce is a requirement to store the bank account and set a flag to
require a verification via network checking.

Test Execution:
------------------------------

Remote:
Finished in 261.78288 seconds.
97 tests, 525 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
   0 notifications
100% passed

Unit:
Finished in 16.583145 seconds.
5045 tests, 74984 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
   0 notifications
100% passed

Rubocop
------------------------------
Inspecting 730 files
730 files inspected, no offenses detected

Closes #4285