Payflow Pro: Add 3DS Support
This commit introduces MPI-compatible way to pass 3DS authentication
parameters, as documented in [Payflow's 3-D Secure with 3rd-Party Merchant Plug-ins
](https://developer.paypal.com/api/nvp-soap/payflow/3d-secure-mpi/).

Prior to this change, we would pass the fields via the `BuyerAuthResult`
inside the `Card` node. Unfortunately, there's no explicit mention of
these, and as far as I can see, this was inferred via a series of
commits, and even then, some of the links referenced in them, are broken
:(, these are the commits:
- #3362
- #4075
- #4113
- #4168

Current Reference documentation, as of March 2022:
- [How to send 3-D Secure authentication data](https://developer.paypal.com/api/nvp-soap/payflow/3d-secure-mpi/#link-payflowfields)

"Legacy" documentation for 3DS data:
- [BuyerAuthStatusEnum, page 109](https://www.paypalobjects.com/webstatic/en_US/developer/docs/pdf/pp_payflowpro_xmlpay_guide.pdf)

ECS-2350

Note: the 10 failing tests are the same ones that fail in `master`.
Note2: additional Rubocop fixes are introduced.:x

Rubocop:
731 files inspected, no offenses detected

Unit:
57 tests, 361 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
39 tests, 169 assertions, 10 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
74.359% passed