StripePI: Fix Invalid Token.
In order to solve the error when attempting to make a purchase using an
Apple Pay payment, with Connect for authentication, because was not include
the options that contain the account information.
This commit includes the options to the tokenization and the request to tokenize
and create intent has the same authentication values.

Local Tests:
---------------------------------------
39 tests, 207 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote Tests:
---------------------------------------
76 tests, 341 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

RuboCop:
---------------------------------------
736 files inspected, no offenses detected