Adyen: Update API version
Summary:
------------------------------
This PR changes the PAYMENT_API_VERSION updating it to a newer version. According to changelog and previous update there
is not any additional requirements to handle the update

Closes #4418

Test Execution:
------------------------------
Unit test
Finished in 0.06828 seconds.
94 tests, 479 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote test
Finished in 163.203064 seconds.
123 tests, 439 assertions, 3 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
97.561% passed
--------------------------
Failures not related with the added code

RuboCop:
------------------------------
739 files inspected, no offenses detected