Wompi: Redirect `refund` to `void`
Refunds issued before 9PM Colombia time will fail. To fix this Wompi is
developing a `refund_or_void` endpoint, which will determine the time and run the proper transaction. To avoid failures in the meantime
all refunds will now be run as `void` calls. This is temporary.

I've created a new unit test to cover the `refund to void` flow and have
commented out the the other refund tests, which are superflous at the
moment.

CE-2635

Unit:
11 tests, 49 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
12 tests, 31 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed