Cybersource: Fixing Amex cryptogram bug
Summary:
------------------------------
Add changes to the 'add_auth_network_tokenization' method so
it is able to handle 20bytes and 40bytes American Express
Network Tokens.

Note: The 3 failing remote tests correspond to authentication
failures because of the lack of a latam related account.

Test Execution:
------------------------------

Remote
Finished in 98.402916 seconds.
115 tests, 590 assertions, 3 failures, 0 errors, 0 pendings,
0 omissions, 0 notifications
97.3913% passed

Unit
Finished in 29.730562 seconds.
5207 tests, 75869 assertions, 0 failures, 0 errors, 0 pendings,
0 omissions, 0 notifications
100% passed

RuboCop
742 files inspected, no offenses detected