CyberSource: updating API version to 1.198 and fixing 3DS test
Summary:
------------------------------
Updates the API version for CyberSource adapter to version 1.198
and fixes a couple of broken tests by updating the pares test
token

Note: The is 1 failing remote test that corresponds to
authentication failures because of the lack of a Latam
related account.

Test Execution:
------------------------------

Remote
Finished in 101.297606 seconds.
115 tests, 594 assertions, 1 failures, 0 errors, 0 pendings,
0 omissions, 0 notifications
99.1304% passed

Unit
Finished in 24.866962 seconds.
5212 tests, 75893 assertions, 0 failures, 0 errors, 0 pendings,
0 omissions, 0 notifications
100% passed

RuboCop
743 files inspected, no offenses detected