Stripe Payment Intents: fix bug with billing address email
This corrects an issue where email was being dropped from billing_details in the request.

The previous assumption was that email would be a field contained within the billing_address hash, but (in Spreedly's integration, at least), it's actually nested at the top-level of the options hash.

This change should not affect behavior of the add_billing_address method for any other integrations with Active Merchant that do supply an email attribute as part of the billing_address hash.

CER-41

LOCAL
5292 tests, 76267 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

746 files inspected, no offenses detected

UNIT
39 tests, 207 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

REMOTE
78 tests, 372 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed