CyberSource: update support for several fields
CER-296

This PR is addressing some issues with some fields not adhering to the order in the XSD file.

The original PR had to be reverted, and in particular I want to call out the `otherTax` element and how network tokenization with Mastercard is causing problems with XML parsing errors. I have added comments to some sections in the XSD to further explain this.

Unit Tests:
124 tests, 592 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote Tests:
122 tests, 625 assertions, 3 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
97.541% passed
*3 failing tests also failing on master

Local Tests:
5400 tests, 76832 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed