Rapyd: Add customer object to transactions
Description
-------------------------
Rapyd Gateway send customer info on store transactions, with this commit
it will be able to send customer on authorize and purchase transaction,
when it use a US “PMT”s include the addresses object into the customer data
in order to be able to perform US transactions properly.

Unit test
-------------------------

Finished in 0.123245 seconds.
22 tests, 103 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote test
-------------------------

Finished in 100.082672 seconds.
33 tests, 92 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed
0.33 tests/s, 0.92 assertions/s

Rubocop
-------------------------

756 files inspected, no offenses detected