Checkout_v2: Add idempotency key support
This PR is to add the support for an optional idempotency key through the header during requests and should be available to all actions 'purchase, authorize, and etc'. I did note that the failing remote tests were sending back 401 unauthorize even when on the latest upstream master. Test results below:
Local:
5469 tests, 77162 assertions, 0 failures, 19 errors, 0 pendings, 0 omissions, 0 notifications
99.6526% passed

Unit:
54 tests, 300 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
89 tests, 213 assertions, 4 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
95.5056% passed