Braintree: Add new stored credential method
ECS-3194

Braintree has informed us we are not handling stored credentials appropriately
in certain cases. This commit adds a new method to handle this new behavior in a
controlled manner by only doing so if a flag (new_stored_credential) indicates we should.

The changes are that
* If it's the initial_transaction & recurring it is to be marked
as `recurring_first`.
* Support for the `installment` reason type
* Map unscheduled to '' transaction_source

Test Summary
Remote:
111 tests, 585 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed