HiPay: Add Gateway
SER-719
SER-772
------

Summary:
------
Adding HiPay gateway with support for basic functionality including authorize, purchase, capture, and store calls.

HiPay requires the tokenization of the CreditCards that's why this first implementation includes the tokenization and store methods.

The store and tokenization methods are different in the "multiuse" param. With this a tokenized PM can be used once(tokenization) or several times(store)

Tests
------
Remote Test:
Finished in 6.627757 seconds.
9 tests, 39 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Unit Tests:
18 tests, 50 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

RuboCop:
787 files inspected, no offenses detected