Rapyd: email mapping update
Description
-------------------------
This commit update the email mapping for purchase,
when we have an customer_id and the payment_method is a token
we will send receipt_email instead email, otherwise the email will be sent
into the customer object .
[SER-1040](https://spreedly.atlassian.net/browse/SER-1040)

Unit test
-------------------------
Finished in 0.186326 seconds.

44 tests, 208 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

236.15 tests/s, 1116.32 assertions/s

Remote test
-------------------------
Finished in 263.428138 seconds.

49 tests, 142 assertions, 1 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
97.9592% passed

0.19 tests/s, 0.54 assertions/s

Rubocop
-------------------------
784 files inspected, no offenses detected