Rapyd: Improve and fix error messages
Description
-------------------------
[SER-1045](https://spreedly.atlassian.net/browse/SER-1045)

This commit update the way to get error codes and sucess transactions.

Rapyd always respond with error_code and response_code fields, and these fields
have the same value now we get both fields in case one of them is empty.

To differentiate when we get a 200 code with an error message from Rapyd,
we have included one validation when the message field is empty to determine if the transaction
is a success transaction or not.

Finally this commit add a small fix when the Rapyd response is not a hash

Unit test
-------------------------
Finished in 0.159334 seconds.

49 tests, 225 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications

100% passed

307.53 tests/s, 1412.13 assertions/s

Remote test
-------------------------
Finished in 234.681137 seconds.

53 tests, 147 assertions, 3 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications

94.3396% passed

0.23 tests/s, 0.63 assertions/s

Rubocop
-------------------------
787 files inspected, 1 offense detected

There is 1 offense not related to this commit