Shift4: Add support for avs_result
CER-1128

This adds support for avs_result in the gateway's response. The standard ActiveMerchant AVS mapping is used, but varies slightly from Shift4's AVS message language in their docs. This should be taken into consideration during review. No remote test was added because the avs hash is not returned in sandbox transactions. It is returned consistently in the same format in the `transaction` hash in production transcripts.

Unit:
28 tests, 176 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
29 tests, 67 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Local:
5814 tests, 79016 assertions, 2 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
99.9656% passed