CyberSource: Extend support for `gratuity_amount`, update Mastercard NT fields
CER-1236

This PR adds the option to send `gratuityAmount` in purchase and auth request. Also updates the order of the request body when Mastercard Network Tokens are used in order to avoid XML parsing errors. `ucaf` parent element is called for MC NT and is not with other card brands. This was causing a conflict/parse error when tax fields are sent as part of the request.

Unit Tests:
148 tests, 818 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote Tests:
135 tests, 668 assertions, 7 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
94.8148% passed
*These 7 tests are also failing on master and in the last update to this gateway

Local Tests:
5827 tests, 79117 assertions, 2 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
99.9657% passed