Orbital: Add support for L2 and L3 fields, update XSD file
A bit of refactoring was needed in the gateway file to support these fields and get things in better shape with adhering to the XSD. Unit tests are now more comprehensive to check schema. I added a remote test and left comments to indicate which fields are restricted to Canadian merchants.

Unit Tests:
148 tests, 850 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote Tests:
134 tests, 504 assertions, 8 failures, 11 errors, 0 pendings, 0 omissions, 0 notifications
85.8209% passed
*Same number of failures and errors on master.

Local Tests:
5899 tests, 79647 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed