Plexo: remove Bin and Last4 fields from NetworkToken
Summary:
------------------------------

For NT transactions we must pass the Bin and the last4
digits of the underlying PAN, otherwise we can't pass it
currently we where passing those values based in the NT number
instead of PAN ones, and due we don't have the underlying PAN,
those lines are unnecesary.

Remote Tests:
------------------------------
Finished in 44.017128 seconds.
32 tests, 65 assertions, 0 failures, 0 errors, 0 pendings, 3 omissions, 0 notifications
100% passed

Unit Tests:
------------------------------
Finished in 0.040727 seconds.
25 tests, 140 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Rubocop
------------------------------
798 files, no offenses detected